pyproject.toml
[project]
name = "discord-linked-role"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask[async]==3.0.0",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "mysql-connector-python==8.2.0",
    "gunicorn==21.2.0",
]
